+++ CHANGELOG.md
@@ -24,6 +24,8 @@
   at `go.uber.org/dig`.
 - **[Breaking]** Pass a tracer the `uhttp/uhttpclient` constructor explicitly, instead
   of using a global tracer. This will allow to use http client in parallel tests.
+- **[Breaking]** Introduce a config loader, this will allow to override config loading
+  and use custom dirs to load from.
 
 ## v1.0.0-beta2 (09 Mar 2017)